+++ VSR/Backend/TF/Framework/Trainer.py
@@ -9,7 +9,6 @@ training methodology for SISR, VSR or other image tasks.
 """
 
 import logging
-import time
 from pathlib import Path
 
 import numpy as np
@@ -91,9 +90,12 @@ class Trainer:
       self._saved.mkdir(parents=True, exist_ok=True)
     if isinstance(self._logd, Path):
       self._logd.mkdir(parents=True, exist_ok=True)
-      if LOG.isEnabledFor(logging.DEBUG):
-        hdl = logging.FileHandler(self._logd / 'training.txt')
-        LOG.addHandler(hdl)
+      _logger = logging.getLogger('VSR')
+      if _logger.isEnabledFor(logging.DEBUG):
+        fd = logging.FileHandler(self._logd / 'vsr_debug.log', encoding='utf-8')
+        fd.setFormatter(
+            logging.Formatter("[%(asctime)s][%(levelname)s] %(message)s"))
+        _logger.addHandler(fd)
     if self.model.compiled:
       self.graph = tf.get_default_graph()
     else:
@@ -260,18 +262,16 @@ class VSR(Trainer):
                                                        shuffle=True,
                                                        memory_limit=mem)
     v.train_loader.prefetch(v.memory_limit)
-    date = time.strftime('%Y-%m-%d %T', time.localtime())
     v.avg_meas = {}
     if v.lr_schedule and callable(v.lr_schedule):
       v.lr = v.lr_schedule(steps=v.global_step)
-    print('| {} | Epoch: {}/{} | LR: {:.2g} |'.format(
-        date, v.epoch, v.epochs, v.lr))
+    LOG.info(f"| Epoch: {v.epoch}/{v.epochs} | LR: {v.lr:.2g} |")
     with tqdm.tqdm(train_iter, unit='batch', ascii=True) as r:
       for items in r:
         self.fn_train_each_step(items)
         r.set_postfix(v.loss)
     for _k, _v in v.avg_meas.items():
-      print('| Epoch average {} = {:.6f} |'.format(_k, np.mean(_v)))
+      LOG.info(f"| Epoch average {_k} = {np.mean(_v):.6f} |")
     if v.epoch % v.validate_every_n_epoch == 0 and v.val_loader:
       self.benchmark(v.val_loader, v, epoch=v.epoch, memory_limit='1GB')
       v.summary_writer.add_summary(self.model.summary(), v.global_step)
@@ -386,6 +386,9 @@ class VSR(Trainer):
     v.mean_metrics = {}
     v.loader = loader
     self.fn_benchmark_body()
+    log_message = str()
     for _k, _v in v.mean_metrics.items():
-      print('{}: {:.6f}'.format(_k, np.mean(_v)), end=', ')
-    print('')
+      _v = np.mean(_v)
+      log_message += f"{_k}: {_v:.6f}, "
+    log_message = log_message[:-2] + "."
+    LOG.info(log_message)